fix: Migrate Home out of redesign
