+++ src/layout/Footer.jsx
@@ -2,7 +2,7 @@ import style from "../style";
 import Section from "../redesign/components/Section";
 import PolicyEngineMainLogo from "../images/logos/policyengine/white.svg";
 import useDisplayCategory from "../hooks/useDisplayCategory";
-import { SubscribeToPolicyEngine } from "../redesign/components/HomeSubscribe";
+import { SubscribeToPolicyEngine } from "../pages/home/HomeSubscribe";
 import {
   FacebookFilled,
   GithubFilled,
