+++ src/pages/home/HomePolicyEngineFeatures.jsx
@@ -1,9 +1,9 @@
 import style from "../../style";
-import Section from "./Section";
+import Section from "../../redesign/components/Section";
 import codingScreenshot from "../../images/home/coding_screenshot.png";
 import parameterScreenshot from "../../images/home/parameter_screenshot.png";
 import decileChartScreenshot from "../../images/home/decile_chart_screenshot.png";
-import ShowcaseItem from "./ShowcaseItem";
+import ShowcaseItem from "../../redesign/components/ShowcaseItem";
 import useCountryId from "../../hooks/useCountryId";
 
 function ImageComponent(imagePath) {
