+++ src/pages/home/HomeQuoteCarousel.jsx
@@ -1,13 +1,13 @@
-import style from "../../style";
+import style from "../../style/index.js";
 import useDisplayCategory from "../../hooks/useDisplayCategory.jsx";
 import { useState, useEffect } from "react";
 import { quoteData } from "../../data/quotes.js";
-import { orgData } from "../../data/organisations";
-import Carousel from "./Carousel";
-import useCountryId from "../../hooks/useCountryId";
+import { orgData } from "../../data/organisations.js";
+import Carousel from "../../redesign/components/Carousel.jsx";
+import useCountryId from "../../hooks/useCountryId.jsx";
 import { Link } from "react-router-dom";
-import FontIcon from "./FontIcon";
-import Section from "./Section";
+import FontIcon from "../../redesign/components/FontIcon.jsx";
+import Section from "../../redesign/components/Section.jsx";
 import ArrowButton from "controls/ArrowButton";
 
 export default function HomeQuoteCarousel() {
