+++ src/pages/home/HomeUsedBy.jsx
@@ -1,5 +1,5 @@
 import style from "../../style";
-import Section from "./Section";
+import Section from "../../redesign/components/Section";
 import { orgData } from "data/organisations";
 import useCountryId from "../../hooks/useCountryId";
 import useDisplayCategory from "../../hooks/useDisplayCategory";
